requirements.txt
torch
transformers
datasets
tqdm
pyyaml
accelerate
peft